feat(imaging): §3 stretch + per-channel histogram, §4 measurement tools
Completes the capture-panel MVP (docs/capture_panel.md §3/§4), display-only.

core/imaging/stretch.py (Qt-free, unit-tested):
- apply_stretch(): black/white + Linear/Log/Asinh + midtones (MTF) → uint8.
- auto_levels() (percentile STF), region_stats(), channel_histograms() (R/G/B
  from the real CFA pixels).
- render_view() now returns LINEAR arrays; stretch is a separate stage.

UI:
- FitsViewer: applies the stretch non-destructively; adds pixel readout on
  hover, a saturation/clipping overlay (full-well from config), and a
  region-stats ROI — all measuring the linear array.
- HistogramDock → "Display" tab: per-channel (R/G/B) histogram, black/white/
  midtones sliders, stretch-mode combo, auto-stretch, saturation + ROI toggles,
  pixel + region readouts, frame stats.
- Image toolbar slimmed to the View selector + "display ≠ data" reminder.
- Acquisition rail gains a Display tab; the log gets the full-width bottom band.

87 tests pass, ruff clean.

+++ tests/core/test_debayer.py
@@ -78,10 +78,10 @@ def test_render_view_shapes_and_dtypes() -> None:
     assert plane.shape == (2, 2) and plane.dtype == np.uint16
 
     sp = d.render_view(arr, d.VIEW_SUPERPIXEL)
-    assert sp.shape == (2, 2, 3) and sp.dtype == np.uint8  # display-normalized
+    assert sp.shape == (2, 2, 3) and sp.dtype == np.uint16  # linear; stretch is separate
 
     interp = d.render_view(arr, d.VIEW_INTERP)
-    assert interp.shape == (4, 4, 3) and interp.dtype == np.uint8
+    assert interp.shape == (4, 4, 3) and interp.dtype == np.uint16
 
 
 def test_render_view_does_not_mutate_raw() -> None:
